% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{atualiza_dados}
\alias{atualiza_dados}
\title{Atualiza os dados da COVID no Brasil.}
\usage{
atualiza_dados()
}
\value{
Base atualizada ate o momento da COVID-19 no Brasil.
}
\description{
Realiza a atualizacao de dados da pandemia de COVID-19 com dados do Portal Brasil.io e do Ministerio da Saude.
}
\details{
ATENCAO: Esta funcao consome algum tempo e uma quantidade razoavel de memoria RAM.
}
