% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{backup_base}
\alias{backup_base}
\title{Faz um backup da ultima atualizacao disponivel}
\usage{
backup_base()
}
\value{
Backup da base mais atualizada disponivel.
}
\description{
Quando invocado, faz uma copia do arquivo final covid para a pasta de dados brutos. Para uso interno no pacote.
}
