% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{base_cidades}
\alias{base_cidades}
\title{Gera a Base Derivada ao nivel de Cidades}
\usage{
base_cidades()
}
\value{
Base Derivada ao nivel de Cidades.
}
\description{
Gera a Base Derivada ao nivel de Cidades com diversos indicadores normalizados para as analises graficas (medias moveis de 7 dias, taxa de mortalidade de casos detectados, taxas por 100 mil habitantes, indicadores normalizados por populacao, indicadores normalizados por area, indicadores normalizados por densidade populacional e indicadores normalizados com logaritmos).
}
