% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{base_estados}
\alias{base_estados}
\title{Gera a Base Derivada ao nivel de Estados}
\usage{
base_estados()
}
\value{
Base Derivada ao nivel de Estados.
}
\description{
Gera a Base Derivada ao nivel de Estados com diversos indicadores normalizados para as analises graficas (medias moveis de 7 dias, taxa de mortalidade de casos detectados, taxas por 100 mil habitantes, indicadores normalizados por populacao, indicadores normalizados por area, indicadores normalizados por densidade populacional e indicadores normalizados com logaritmos).
}
