% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{bases_derivadas}
\alias{bases_derivadas}
\title{Gera as Bases de Dados Derivadas para a geracao de graficos.}
\usage{
bases_derivadas()
}
\value{
Bases derivadas com sumarizacoes e indicadores normalizados nos niveis de Cidades, Regioes de Saude, Estados, Regioes Politicas e Brasil inteiro.
}
\description{
A partir da para de dados atualizada com dados da pandemia de COVID-19 (com dados do Portal Brasil.io e do Ministerio da Saude), gera as versoes derivadas aos niveis de Cidades, Regioes de Saude, Estados, Regioes Politicas e Brasil inteiro. Indicadores normalizados tambem sao acrescentados para um leque de analises.
}
