% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{carrega_auxiliares}
\alias{carrega_auxiliares}
\title{Carrega dados auxiliares.}
\usage{
carrega_auxiliares(arquivo)
}
\arguments{
\item{arquivo}{Base de Dados a ser carregada.}
}
\value{
Base auxiliar processada.
}
\description{
Carrega e processa dados auxiliares para uso na montagem da base posteriormente.
}
