% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{deriva_codigo_municipio}
\alias{deriva_codigo_municipio}
\title{Deriva dados auxiliares relacionados ao Codigio de Municipio nao-oficial}
\usage{
deriva_codigo_municipio()
}
\value{
Dados auxiliares relacionados ao Codigio de Municipio.
}
\description{
Deriva dados auxiliares relacionados ao Codigio de Municipio nao-oficial
}
