% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infos_geograficas}
\alias{infos_geograficas}
\title{Informacoes Geograficas Auxiliares.}
\format{
Uma tibble com 5.572 linhas e 5 colunas:
\describe{
\item{cod_ibge}{Codigo de Municipios do IBGE (Instituto Brasileiro de Geografia e Estatistica)}
\item{lat}{Latitude do Centroide do Municipio}
\item{lon}{Longitude do Centroide do Municipio}
\item{area_mun_km2}{Area do Municipio em Km2}
\item{capital}{Indicador binario do Municipio ser capital do estado}
}
}
\source{
\url{Bruno Cesar Pasquini}
}
\usage{
infos_geograficas
}
\description{
Um Banco de Dados Geograficos auxiliares contendo informacoes adicionais para uso nas analises.
}
\keyword{datasets}
