% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{le_brasil_io}
\alias{le_brasil_io}
\title{Importa dados da COVID do Portal Brasil.io}
\usage{
le_brasil_io()
}
\value{
Base importada e tratada da COVID do Portal Brasil.io.
}
\description{
Realiza a importacao e o tratamento dos dados da COVID-19 do Portal Brasil.io.
}
