% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{le_ministerio}
\alias{le_ministerio}
\title{Importa dados da COVID do Ministerio da Saude}
\usage{
le_ministerio()
}
\value{
Base importada e tratada da COVID do Ministerio da Saude.
}
\description{
Realiza a importacao e o tratamento dos dados da COVID-19 do Ministerio da Saude (ou de seu backup).
}
\details{
ATENCAO: Esta sub-funcao consome muita memoria e leva algum tempo. Por favor aguarde...
}
