% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{semana_epid}
\alias{semana_epid}
\title{Informacoes Auxiliares sobre as Semanas Epidemiologicas.}
\format{
Uma tibble com 365 linhas e 2 colunas:
\describe{
\item{date}{Data}
\item{semana_epidem}{Semana Epidemiologica correspondente a data}
}
}
\source{
\url{https://opendatasus.saude.gov.br/dataset/bd-srag-2020}
}
\usage{
semana_epid
}
\description{
Um Banco de Dados auxiliares sobre as primeiras Semanas Epidemiologicas para uso nas analises. Este banco de dados cobre as primeiras 77 semanas epidemiologicas, uma vez que as bases de dados atualizadas contem informacao somente a partir da semana epidemiologica 9.
}
\keyword{datasets}
