% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{sumarios_derivacoes}
\alias{sumarios_derivacoes}
\title{Processas as Sumarizacoes de Area e Populacao}
\usage{
sumarios_derivacoes()
}
\value{
Arquivos com as sumarizacoes de Area e Populacao 2019.
}
\description{
Processa externamente as sumarizacoes de Area (em Km2) e de Populacao (conforme apurado pelo TCU no ano de 2019)
}
