% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabela_mun}
\alias{tabela_mun}
\title{Tabela com Mapa dos Municipios do Brasil.}
\format{
Uma tibble com 5.572 linhas e 8 colunas:
\describe{
\item{code_muni}{Codigo do Municipio (de acordo com o IBGE)}
\item{name_muni}{Nome do Municipio}
\item{code_state}{Codigo do Estado (de acordo com o IBGE)}
\item{abbrev_state}{Sigla do Estado (Unidade Federativa)}
\item{name_state}{Nome do Estado}
\item{code_region}{Codigo da Regiao}
\item{name_region}{Nome da Regiao}
\item{geom}{Poligono com a Geometria do Estado}
}
}
\source{
\url{https://github.com/ipeaGIT/geobr}
}
\usage{
tabela_mun
}
\description{
Um Banco de Dados auxiliares com informacoes geograficas e cartograficas dos municipios do Brasil. Os dados aparentemente sao oficiais, vindos do IBGE (Instituto Brasileiro de Geografia e Estatistica).
}
\keyword{datasets}
