% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabela_ufs}
\alias{tabela_ufs}
\title{Tabela com Mapa dos Estados do Brasil.}
\format{
Uma tibble com 27 linhas e 6 colunas:
\describe{
\item{code_state}{Codigo do Estado (de acordo com o IBGE)}
\item{abbrev_state}{Sigla do Estado (Unidade Federativa)}
\item{name_state}{Nome do Estado}
\item{code_region}{Codigo da Regiao (de acordo com o IBGE)}
\item{name_region}{Nome da Regiao}
\item{geom}{Poligono com a Geometria do Estado}
}
}
\source{
\url{https://github.com/ipeaGIT/geobr}
}
\usage{
tabela_ufs
}
\description{
Um Banco de Dados auxiliares com informacoes geograficas e cartograficas dos estados do Brasil. Os dados aparentemente sao oficiais, vindos do IBGE (Instituto Brasileiro de Geografia e Estatistica).
}
\keyword{datasets}
