% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncoes.R
\name{tema_bruno}
\alias{tema_bruno}
\title{Definicoes graficas para graficos do R}
\usage{
tema_bruno()
}
\value{
Um tema definido para construcao de graficos.
}
\description{
Definicoes de elementos graficos para construcao de analises graficas.
}
